rename the new fields + move nested date fields
