+++ packages/editor/src/dataviews/store/private-actions.ts
@@ -43,9 +43,9 @@ import {
 	stickyField,
 	descriptionField,
 	readOnlyDescriptionField,
-	postsPerPageField,
-	siteDiscussionField,
-	postsPageTitleField,
+	templatePostsPerPageField,
+	templateSiteDiscussionField,
+	templatePostsPageTitleField,
 } from '@wordpress/fields';
 import {
 	altTextField,
@@ -304,9 +304,12 @@ export const registerPostTypeSchema =
 				// target other entities (`root/site` and the posts page). They
 				// are merged into the form data under a namespace and routed
 				// back to their entity by `DataFormPostSummary`.
-				postTypeConfig.slug === 'wp_template' && postsPageTitleField,
-				postTypeConfig.slug === 'wp_template' && postsPerPageField,
-				postTypeConfig.slug === 'wp_template' && siteDiscussionField,
+				postTypeConfig.slug === 'wp_template' &&
+					templatePostsPageTitleField,
+				postTypeConfig.slug === 'wp_template' &&
+					templatePostsPerPageField,
+				postTypeConfig.slug === 'wp_template' &&
+					templateSiteDiscussionField,
 				postTypeConfig.supports?.editor &&
 					postTypeConfig.viewable &&
 					postPreviewField,
