+++ packages/fields/README.md
@@ -168,14 +168,6 @@ Ping status field for BasePost.
 
 Post content information field for BasePost.
 
-### postsPageTitleField
-
-Posts page title field for the `home`/`index` template summary.
-
-### postsPerPageField
-
-Posts per page field for the `home`/`index` template summary.
-
 ### PostType
 
 Undocumented declaration.
@@ -208,10 +200,6 @@ Restore action for PostWithPermissions.
 
 ScheduledDate Field.
 
-### siteDiscussionField
-
-Site discussion field for the `home`/`index` template summary.
-
 ### slugField
 
 Slug field for BasePost.
@@ -227,6 +215,18 @@ Sticky field for BasePost.
 ### templateField
 
 Template field for BasePost.
+
+### templatePostsPageTitleField
+
+Posts page title field for the `home`/`index` template summary.
+
+### templatePostsPerPageField
+
+Posts per page field for the `home`/`index` template summary.
+
+### templateSiteDiscussionField
+
+Discussion field for the `home`/`index` template summary.
 
 ### templateTitleField
